ccms-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
# psycopg2-binary removed — use psycopg3 which has pre-built wheels for Python 3.13
psycopg[binary]==3.2.4
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
alembic==1.13.2
